### 2019-04-05 #1365 @whedon
[PRE REVIEW]: hankel: A Python library for performing simple and accurate Hankel transformations
**Submitting author:** @steven-murray (<a href="http://orcid.org/0000-0003-3059-3823">Steven Murray</a>)
**Repository:** <a href="https://github.com/steven-murray/hankel" target ="_blank">https://github.com/steven-murray/hankel</a>
**Version:** v0.3
