### 2019-04-08 #1365 @steven-murray
I am wondering if the PDF failed to compile because the paper.md is not on the master branch. Is there a way to point to an alternate branch, or must we point to the master branch?
