### 2019-04-10 #1365 @arfon
I've updated the Git repo address to `https://github.com/steven-murray/hankel`
