### 2019-04-12 #1365 @melissawm
Hello,

It will be my pleasure! Count me in.

Best regards,

Melissa

Em sex, 12 de abr de 2019 08:11, Juanjo Bazán <notifications@github.com>
escreveu:

> 👋 @melissawm <https://github.com/melissawm> and @drvinceknight
> <https://github.com/drvinceknight> : Would you be willing to review this
> paper? I think your expertise in Math and Python makes you a great fit for
> this submission.
>
> I'm willing to start the review if one of you are available, but it would
> be great to have you both.
>
