### 2019-04-22 #1397 @xuanxu
Git repo address updated from `https://github.com/steven-murray/hankel/tree/joss-paper` to `https://github.com/steven-murray/hankel` so whedon can clone it.
