### 2019-04-22 #1397 @xuanxu
@arfon could you do your magic so we can get the pdf paper? (it compiled in the [pre-review issue](https://github.com/openjournals/joss-reviews/issues/1365) but I can replicate it here)
