### 2019-05-12 #1397 @labarba
OK, @melissawm — the proof has compiled. Please start your review. Feel free to ask questions here, and to open issues in the submission repository, as needed. 🙏
