### 2019-05-20 0f83949
fixes #23
Added DOI for Ogata and Szapudi both in readme and joss paper .bib
