+++ paper/paper.bib
@@ -6,6 +6,7 @@
   author = {Ogata, Hidenori},
   year = {2005},
   pages = {949-970},
+  doi = 10.2977/prims/1145474602,
 }
 
 @article{Szapudi2005,
@@ -17,6 +18,7 @@
   month = may,
   year = {2005},
   pages = {4-4},
+  doi = 10.1086/496971
 }
 @incollection{Huybrechs2008,
   title = {Highly {{Oscillatory Quadrature}}},
