### 2019-05-20 2ac3dc2
Add "how to run tests" to README
Suggestion from #24.
