### 2019-05-20 e572b04
contributing guidelines
From suggestion #25. Removes the need for #24 as well.
