### 2019-05-21 #1397 @melissawm
Hello, @xuanxu , maybe I didn't understand the workflow for the review process. I'm doing the review, however there are some issues that I'm waiting to be solved before I can check all of the boxes and finish the review. I've opened these issues in the source repository as instructed. Should I not wait for that?
