### 2019-05-21 #1397 @xuanxu
Hi @melissawm! I didn't see the issues you opened in the hankel repository, sorry. The workflow you are following is OK.
Just a detail: as the central place for the review process, it'd be ideal if you could add a short comment in this issue whenever the status changes or to keep track of the actions being taken (for example, something like "I've opened some issues in the target repo with a couple problems that need to be addressed" in this case). That way this issue works as a log of the review.
