### 2019-05-23 #1397 @melissawm
Hello all. @xuanxu the review has been finalized. @steven-murray has answered to all the issues that I've opened (adding the DOI for the references; adding a document/guide for contributors and those seeking support, including how to test the software). As far as I'm concerned this is approved.
