### 2019-05-23 #1397 @xuanxu
@steven-murray can you merge the changes into master?
