### 2019-05-24 #1397 @xuanxu
@steven-murray the paper compilation failed because there's a couple of missing braces. I've issued a [PR in your repo](https://github.com/steven-murray/hankel/pull/26) to fix that.
