### 2019-05-25 #1397 @xuanxu
@steven-murray Whedon suggests `https://doi.org/10.1017/cbo9781139107136.003` as a posible DOI for the `Highly Oscillatory Quadrature` citation. Do you think it's the right one? If so I have issued a PR in the hankel repo updating it.
