### 2019-05-28 #1397 @steven-murray
Yes, that seems right, I've merged it, thanks!
