### 2019-05-28 #1397 @xuanxu
`The source code for hankel has been archived to Zenodo with the linked DOI: [@zenodo]`

The Zenodo archive will be linked on the first page of the paper, so we ask that you don't add an additional link in the text or references. Can you edit that citation out?
