### 2019-05-28 944df2f
remove zenodo self-reference from paper
