+++ paper/paper.bib
@@ -54,12 +54,3 @@
   year={2001},
   publisher={Springer}
 }
-@article{zenodo,
-  	Author = {Steven G. Murray and Francis J. Poulin and Sebastian Müller and David Zaslavsky},
-  	Doi = {10.5281/zenodo.1336791},
-  	Month = {Aug},
-  	Publisher = {Zenodo},
-  	Title = {hankel},
-  	Year = {2018},
-  	url = {https://doi.org/10.5281/zenodo.1336791}
-}
