### 2019-05-28 #1397 @xuanxu
> I have removed it now, thanks.

It is still in the paper and bib files, maybe forgot to push?
