### 2019-05-28 #1397 @steven-murray
Hmm, this seems to have been a line added in the wrong place in the
previous PR. Fixed now!

On Tue, May 28, 2019 at 1:53 PM Juanjo Bazán <notifications@github.com>
wrote:

> From the summary:
> In this single sentence:
>
> Highly oscillatory integrals are a topic of much interest in applied
> mathematics, and there does not exist a general optimal solution to
> numerically evaluate them double-exponential variable transformation based
> on the zeros of the Bessel function [@Ooura1999] has the property that the
> numerical integral converges with many fewer divisions compared to naively
> computing the transform integral.
>
> I guess there's a missing dot somewhere.
>
