### 2019-05-28 #1397 @xuanxu
Thanks @steven-murray, the paper looks ok to me now.

Please release a new version. It will include all the changes made during the review.
Then create a Zenodo archive of it and report the DOI here (make sure to edit the metadata so title and author are the same as in the paper).
