### 2019-05-30 #1397 @steven-murray
Thanks @xuanxu .

I'm not too familiar with Zenodo. I already have a zenodo entry for hankel (https://zenodo.org/record/1336792). It looks like I have three options: create a release, edit the version that's already there, or create a new version. It's not clear to me which path I should take. The latter two are giving me internal server errors at the moment...
