### 2019-05-30 #1397 @xuanxu
@steven-murray I suggest you try these steps:
- Create a new release in Github: https://github.com/steven-murray/hankel/releases
- If Zenodo has enough permissions it should automatically upload the new zip file, otherwise it shuold be done manually: https://zenodo.org/account/settings/github/
- Edit the metadata of the new entry in Zenodo
