### 2019-05-30 #1397 @xuanxu
@steven-murray great! Please edit the title to match the paper:
`hankel: A Python library for performing simple and accurate Hankel transformations`
You can do so here: https://zenodo.org/deposit/3235680
