### 2019-05-31 #1397 @steven-murray
OK cool, done that.

On Thu, May 30, 2019 at 1:09 PM Juanjo Bazán <notifications@github.com>
wrote:

> @steven-murray <https://github.com/steven-murray> great! Please edit the
> title to match the paper:
> hankel: A Python library for performing simple and accurate Hankel
> transformations
> You can do so here: https://zenodo.org/deposit/3235680
> <https://secure-web.cisco.com/1F7TYK4WmghACn4SBF_FTXPWI2ZxZXa3QGa1q2IMmm6YHObR47unS4jTzc0L2RmVz-QjDDHtv4McUQaFV-LNwuQsDXeAAEBfZT1X1KZb5SzqG33QoGLoMuQGVx6mCNRSK6oZi65aoKlfYssKOjevFchnVb_SC_Q7NaxrM0Qy_5gBA3ryXLTLrHcOcH25HD1R03TiEM-Oj0280r6tJie_lUKdSvZ0IFRK3GoL6Wp-2DmwaXYhWtZmtQj47-0unVgqSIdBf11rhI0yojxNVFICVaNK8P3_nHP6dIKTmFxqKmFygBIVL65X-4iPOgFnvPKVAMysA8EmSEKQGupfOB4hfm48IINvo9Y6ijwRgj-UruNss_aH-faXRZX4K2iB6Yr7hEfM_wVNQrTyO8v9JfrJaAIWAScVKAHW8jYYGQ3pHS1FPXBDDAw1DDq-3F5jxve9s0Uq9ivJzB3hJ9NnMy1S1M_loPT4T5ODXbwKpjHWOuEs/https%3A%2F%2Fzenodo.org%2Fdeposit%2F3235680>
>
