### 2019-05-31 #1397 @kyleniemeyer
Hi @steven-murray, nearly ready to accept—I just submitted a small PR to your paper which fixes the formatting of some in-text references: https://github.com/steven-murray/hankel/pull/28
