### 2019-05-31 #1397 @steven-murray
Awesome, thanks! That's now merged.

On Fri, May 31, 2019 at 2:44 PM Kyle Niemeyer <notifications@github.com>
wrote:

> Hi @steven-murray <https://github.com/steven-murray>, nearly ready to
> accept—I just submitted a small PR to your paper which fixes the formatting
> of some in-text references: steven-murray/hankel#28
> <https://github.com/steven-murray/hankel/pull/28>
>
