### 2019-05-31 #1397 @kyleniemeyer
Congrats @steven-murray on your paper's acceptance in JOSS! Thanks to @melissawm for reviewing and @xuanxu for editing!
